Use template to define base java nio buffer types
