+++ javalib/src/main/scala/java/nio/Buffers.scala.gyb
@@ -1,0 +1,303 @@
+// format: off
+package java.nio
+
+// Ported from Scala.js
+import scala.scalanative.unsafe
+%{
+   variants = [
+    ('Byte',   '-547316498' ),
+    ('Char',   '-182887236' ),
+    ('Short',  '383731478'  ),
+    ('Int',    '39599817'   ),
+    ('Long',   '-1709696158'),
+    ('Float',  '1920204022' ),
+    ('Double', '2140173175' )
+   ]
+}%
+% for (T, Seed) in variants:
+object ${T}Buffer {
+  private final val HashSeed = ${Seed} // "java.nio.${T}Buffer".##
+
+  def allocate(capacity: Int): ${T}Buffer = wrap(new Array[${T}](capacity))
+
+% if T == 'Byte':
+  def allocateDirect(capacity: Int): ${T}Buffer = allocate(capacity)
+% end
+
+  def wrap(array: Array[${T}], offset: Int, length: Int): ${T}Buffer =
+    Heap${T}Buffer.wrap(array, 0, array.length, offset, length, false)
+
+  def wrap(array: Array[${T}]): ${T}Buffer =  wrap(array, 0, array.length)
+
+% if T == 'Char':
+  def wrap(csq: CharSequence, start: Int, end: Int): CharBuffer =
+    StringCharBuffer.wrap(csq, 0, csq.length(), start, end - start)
+
+  def wrap(csq: CharSequence): CharBuffer = wrap(csq, 0, csq.length())
+% end
+
+    // Extended API
+% if T == 'Byte':
+  def wrapPointer${T}(array: unsafe.Ptr[${T}], length: Int): ${T}Buffer =
+    Pointer${T}Buffer.wrap(array, length)
+% end
+}
+
+abstract class ${T}Buffer private[nio] (
+    _capacity: Int,
+    override private[nio] val _array: Array[${T}],
+    private[nio] val _offset: Int
+) extends Buffer(_capacity)
+    with Comparable[${T}Buffer] 
+% if T == 'Char':
+    with CharSequence
+    with Appendable
+    with Readable
+%end
+  {
+  private[nio] type ElementType = ${T}
+  private[nio] type BufferType = ${T}Buffer
+
+%if T == 'Byte':
+  private[nio] var _isBigEndian: Boolean = true
+% end
+
+  private def genBuffer = GenBuffer[${T}Buffer](this)
+
+  def this(_capacity: Int) = this(_capacity, null: Array[${T}], -1)
+
+  def slice(): ${T}Buffer
+
+  def duplicate(): ${T}Buffer
+
+  def asReadOnlyBuffer(): ${T}Buffer
+
+  def get(): ${T}
+
+  def put(b: ${T}): ${T}Buffer
+
+  def get(index: Int): ${T}
+
+  def put(index: Int, b: ${T}): ${T}Buffer
+
+  @noinline
+  def get(dst: Array[${T}], offset: Int, length: Int): ${T}Buffer =
+    genBuffer.generic_get(dst, offset, length)
+
+  def get(dst: Array[${T}]): ${T}Buffer =
+    get(dst, 0, dst.length)
+
+  @noinline
+  def put(src: ${T}Buffer): ${T}Buffer =
+    genBuffer.generic_put(src)
+
+  @noinline
+  def put(src: Array[${T}], offset: Int, length: Int): ${T}Buffer =
+    genBuffer.generic_put(src, offset, length)
+
+  final def put(src: Array[${T}]): ${T}Buffer =
+    put(src, 0, src.length)
+    
+% if T == 'Char':
+  def put(src: String, start: Int, end: Int): CharBuffer =
+    put(CharBuffer.wrap(src, start, end))
+
+  final def put(src: String): CharBuffer =
+    put(src, 0, src.length)
+% end 
+
+  @inline final def hasArray(): Boolean =
+    genBuffer.generic_hasArray()
+
+  @inline final def array(): Array[${T}] =
+    genBuffer.generic_array()
+
+  @inline final def arrayOffset(): Int =
+    genBuffer.generic_offset()
+
+  @inline override def position(newPosition: Int): ${T}Buffer = {
+    super.position(newPosition)
+    this
+  }
+
+  @inline override def limit(newLimit: Int): ${T}Buffer = {
+    super.limit(newLimit)
+    this
+  }
+
+  @inline override def mark(): ${T}Buffer = {
+    super.mark()
+    this
+  }
+
+  @inline override def reset(): ${T}Buffer = {
+    super.reset()
+    this
+  }
+
+  @inline override def clear(): ${T}Buffer = {
+    super.clear()
+    this
+  }
+
+  @inline override def flip(): ${T}Buffer = {
+    super.flip()
+    this
+  }
+
+  @inline override def rewind(): ${T}Buffer = {
+    super.rewind()
+    this
+  }
+
+  def compact(): ${T}Buffer
+
+  def isDirect(): Boolean
+
+  // toString(): String inherited from Buffer
+
+  @noinline
+  override def hashCode(): Int =
+    genBuffer.generic_hashCode(${T}Buffer.HashSeed)
+
+  override def equals(that: Any): Boolean = that match {
+    case that: ${T}Buffer => compareTo(that) == 0
+    case _                => false
+  }
+
+  @noinline
+  def compareTo(that: ${T}Buffer): Int =
+    genBuffer.generic_compareTo(that)(_.compareTo(_))
+
+%if T == 'Byte':
+  final def order(): ByteOrder =
+    if (_isBigEndian) ByteOrder.BIG_ENDIAN
+    else ByteOrder.LITTLE_ENDIAN
+  final def order(bo: ByteOrder): ByteBuffer = {
+    if (bo == null)
+      throw new NullPointerException
+    _isBigEndian = bo == ByteOrder.BIG_ENDIAN
+    this
+  }
+%else:
+  def order(): ByteOrder
+%end 
+
+%if T == 'Byte':
+  def getChar(): Char
+  def putChar(value: Char): ByteBuffer
+  def getChar(index: Int): Char
+  def putChar(index: Int, value: Char): ByteBuffer
+
+  def asCharBuffer(): CharBuffer
+
+  def getShort(): Short
+  def putShort(value: Short): ByteBuffer
+  def getShort(index: Int): Short
+  def putShort(index: Int, value: Short): ByteBuffer
+
+  def asShortBuffer(): ShortBuffer
+
+  def getInt(): Int
+  def putInt(value: Int): ByteBuffer
+  def getInt(index: Int): Int
+  def putInt(index: Int, value: Int): ByteBuffer
+
+  def asIntBuffer(): IntBuffer
+
+  def getLong(): Long
+  def putLong(value: Long): ByteBuffer
+  def getLong(index: Int): Long
+  def putLong(index: Int, value: Long): ByteBuffer
+
+  def asLongBuffer(): LongBuffer
+
+  def getFloat(): Float
+  def putFloat(value: Float): ByteBuffer
+  def getFloat(index: Int): Float
+  def putFloat(index: Int, value: Float): ByteBuffer
+
+  def asFloatBuffer(): FloatBuffer
+
+  def getDouble(): Double
+  def putDouble(value: Double): ByteBuffer
+  def getDouble(index: Int): Double
+  def putDouble(index: Int, value: Double): ByteBuffer
+
+  def asDoubleBuffer(): DoubleBuffer
+
+% elif T == 'Char':
+  override def toString(): String = {
+    if (_array != null) {
+      // even if read-only
+      new String(_array, position() + _offset, remaining())
+    } else {
+      val chars = new Array[Char](remaining())
+      val savedPos = position()
+      get(chars)
+      position(savedPos)
+      new String(chars)
+    }
+  }
+
+  final def length(): Int = remaining()
+
+  final def charAt(index: Int): Char = get(position() + index)
+
+  def subSequence(start: Int, end: Int): CharSequence
+
+  def append(csq: CharSequence): CharBuffer =
+    put(csq.toString())
+
+  def append(csq: CharSequence, start: Int, end: Int): CharBuffer =
+    put(csq.subSequence(start, end).toString())
+
+  def append(c: Char): CharBuffer =
+    put(c)
+
+  def read(target: CharBuffer): Int = {
+    // Attention: this method must not change this buffer's position
+    val n = remaining()
+    if (n == 0) -1
+    else if (_array != null) {
+      // even if read-only
+      genBuffer.generic_put(_array, _offset, n)
+      n
+    } else {
+      val savedPos = position()
+      target.put(this)
+      position(savedPos)
+      n
+    }
+  }
+% end
+
+  // Internal API
+% if T == 'Byte':
+  override private[nio] def isBigEndian: Boolean = _isBigEndian
+% end
+
+  private[nio] def load(index: Int): ${T}
+
+  private[nio] def store(index: Int, elem: ${T}): Unit
+
+  @inline
+  private[nio] def load(
+      startIndex: Int,
+      dst: Array[${T}],
+      offset: Int,
+      length: Int
+  ): Unit =
+    genBuffer.generic_load(startIndex, dst, offset, length)
+
+  @inline
+  private[nio] def store(
+      startIndex: Int,
+      src: Array[${T}],
+      offset: Int,
+      length: Int
+  ): Unit =
+    genBuffer.generic_store(startIndex, src, offset, length)
+}
+
+% end
