+++ scripts/gyb_all.sh
@@ -39,6 +39,7 @@ gyb_files $nativelib $scala $runtime Arrays Boxes Primitives
 gyb clib/src/main/scala/scala/scalanative/libc/stdatomic.scala.gyb
 gyb clib/src/main/resources/scala-native/stdatomic.c.gyb
 
+gyb $javaNIO/Buffers.scala.gyb
 gyb $javaNIO/HeapBuffers.scala.gyb
 gyb $javaNIO/HeapByteBufferViews.scala.gyb
 gyb $javaNIO/MappedByteBufferViews.scala.gyb
